+++ src/material/button-toggle/testing/BUILD.bazel
@@ -21,31 +21,12 @@ filegroup(
 )
 
 ng_test_library(
-    name = "harness_tests_lib",
-    srcs = [
-        "button-toggle-group-shared.spec.ts",
-        "button-toggle-shared.spec.ts",
-    ],
+    name = "unit_tests_lib",
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/testing",
         "//src/cdk/testing/testbed",
-        "//src/material/button-toggle",
-    ],
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = [
-            "button-toggle-shared.spec.ts",
-            "button-toggle-group-shared.spec.ts",
-        ],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
         "//src/material/button-toggle",
     ],
 )
